remove collection module
