+++ tests/integrations/langchain_/test_langchain_with_openai_chat_models.py
@@ -3,11 +3,10 @@ import uuid
 
 import pytest
 from langchain.chains import LLMChain
-from langchain.llms.openai import OpenAI
+from langchain.chat_models.openai import ChatOpenAI
 from langchain.prompts import PromptTemplate
 
 from greptimeai.langchain.callback import GreptimeCallbackHandler
-from greptimeai.openai_patcher import _collector
 
 from ..database.db import get_trace_data, truncate_tables
 
@@ -18,19 +17,19 @@ def _truncate_tables():
     yield
 
 
-def test_llm(_truncate_tables):
+def test_chat(_truncate_tables):
     user_id = str(uuid.uuid4())
     model = "gpt-3.5-turbo"
 
     callback = GreptimeCallbackHandler()
-    llm = OpenAI(model=model)
+    chat = ChatOpenAI(model=model)
     prompt = PromptTemplate.from_template("1 + {number} = ")
 
-    chain = LLMChain(llm=llm, prompt=prompt, callbacks=[callback])
+    chain = LLMChain(llm=chat, prompt=prompt, callbacks=[callback])
     result = chain.run(number=1, callbacks=[callback], metadata={"user_id": user_id})
     print(f"{result=}")
 
-    _collector._collector._force_flush()
+    callback.collector._force_flush()
 
     trace = get_trace_data(user_id=user_id, span_name="langchain_llm")
     retry = 0
