+++ tests/integrations/openai_tracker/test_sync.py
@@ -31,7 +31,7 @@ def test_chat_completion(_truncate_tables):
     )
     assert resp.choices[0].message.content == "2"
 
-    _collector._collector._force_flush()
+    _collector._force_flush()
 
     trace = get_trace_data(user_id)
     retry = 0
